# iroha-rest-server-swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a REST server to interact with Hyperledger Iroha Blockchain",
        "version": "1.0.0",
        "title": "Iroha REST Server",
        "contact": {
            "email": "dominic.aidoo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http","https"],
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/v1/asset/createAsset": {
            "post" : {
                "summary" : "Create an asset",
                "description": "Create an asset",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "tags":["AssetController"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Create an asset",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction committed",
                        "schema": {
                            "$ref": "#/definitions/IrohaSuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Error from iroha",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/asset/addAssetQuantity": {
            "post" : {
                "summary" : "Add asset quantity",
                "description": "Add asset quantity",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "tags":["AssetController"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Add asset quantity",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AdjustAssetQuantityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                          "$ref": "#/definitions/IrohaSuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/asset/subtractAssetQuantity": {
            "post" : {
                "summary" : "Subtract asset quantity",
                "description": "Subtract asset quantity",
                "produces": ["application/json"],
                "tags":["AssetController"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Subtract asset quantity",
                        "required": true,
                        "schema": {
                          "$ref": "#/definitions/AdjustAssetQuantityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/IrohaSuccessResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/asset/transferAsset": {
            "post" : {
                "summary" : "Transfer asset",
                "description": "Transfer asset",
                "produces": ["application/json"],
                "tags":["AssetController"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Transfer asset",
                        "required": true,
                        "schema": {
                          "$ref": "#/definitions/TransferAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/IrohaSuccessResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/grantableperm/grantPermission": {
            "post" : {
                "summary" : "Grant a grantable permission",
                "description": "Grant a grantable permission",
                "produces": ["application/json"],
                "tags":["GrantablePermissionController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Grant a grantable permission",
                        "required": true,
                        "schema":{
                            "$ref": "#/definitions/GrantablePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/grantableperm/revokePermission": {
            "post" : {
                "summary" : "Revoke a grantable permission",
                "description": "Revoke a grantable permission",
                "produces": ["application/json"],
                "tags":["GrantablePermissionController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Revoke a grantable permission",
                        "required": true,
                        "schema":{
                            "$ref": "#/definitions/GrantablePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/signatory/addSignatory": {
            "post" : {
                "summary" : "Add signing key to account",
                "description": "Add signing key to account",
                "produces": ["application/json"],
                "tags":["SignatoryController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Add signing key to account",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SignatoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/signatory/removeSignatory": {
            "post" : {
                "summary" : "Remove signing key to account",
                "description": "Remove signing key to account",
                "produces": ["application/json"],
                "tags":["SignatoryController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Remove signing key to account",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SignatoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
...
    }, 
    "definitions": {
        "CreateAssetRequest": {
            "type": "object",
            "properties": {
                "assetName": {
                     "type": "string"
                },
                "domainId": {
                    "type": "string"
                },
                "precision": {
                    "type": "integer"
                }
            }
        },
        "AdjustAssetQuantityRequest": {
            "type": "object",
            "properties": {
                "assetId": {
                     "type": "string"
                },
                "amount": {
                    "type": "string"
                }
            }
        },
        "TransferAssetRequest": {
            "type": "object",
            "properties": {
                "srcAccountId": {
                     "type": "string"
                },
                "destAccountId": {
                     "type": "string"
                },
                "assetId": {
                     "type": "string"
                },
                "description": {
                     "type": "string"
                },
                "amount": {
                    "type": "string"
                }
            }
        },
...
        "GrantablePermissionRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "permission": {
                     "type": "string"
                }
            }
        },
...
        "SignatoryRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "publicKey": {
                    "type": "string"
               }
            }
        },
...
        "getAssetInfoResponse": {
            "type": "object",
            "properties": {
                "assetId": {
                     "type": "string"
                },
                "domainId": {
                     "type": "string"
                },
                "precision": {
                    "type": "number"
                }
            }
        },
        "IrohaSuccessResponse": {
            "type": "object",
            "properties": {
                "txHash": {
                     "type": "string",
                     "example": "5a33387b7e0cfe4b2d79fdc9bbba61a235a268eef211a73a282b9ce30ac06e28"
                },
                "status": {
                    "type": "string",
                    "example": "COMMITTED"
                }
            }
        },
        "IrohaErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                     "type": "string",
                     "example": "Command response error: expected=COMMITTED, actual=REJECTED"
                },
                "status": {
                    "type": "string",
                    "example": "REJECTED"
                }
            }
        },
...
    }
}